Fix NPE in FSGrad
When a segment is passing the neighbour tile border twice and the tap is in the first part, then points from other tile are excluded, but for the first point in the third part last.getNeighbour(node) returns null. Solution is to limit the segmentNodes during calculation to one tile.